+++ CHANGES.md
@@ -1,6 +1,7 @@
 
 # Unreleased
 
+* fix `utils.get_array_statistics` method to avoid `ZeroDivisionError` when there is no valid pixel
 * use `GDAL_MEM_ENABLE_OPEN=TRUE` when opening a numpy array with rasterio
 
 # 7.5.0 (2025-02-26)